readme love
1. This is my penultimate editing workover of the readme.txt in order:
- to pass the http://wordpress.org/plugins/about/validator/,
- to bring editorial consistency to the file as a whole,
- to add a few links where appropriate,
- and to try to distinguish the Clef service in general and the WP Clef
plugin specifically.

In some of my wording choices I'm likely overstepping my bounds: the
Clef team needs to choose how it wants to "market" this plugin and
whether it makes sense to distinguish Clef and WP Clef.

2. There are a couple out-of-date items that would be nice to update
before pushing a plugin update: the Vimeo video still shows the old QR
code instead of the new (and enchantingly beautiful!) Wave; the
screenshots need to be updated to reflect the new settings.

3. Removed an extraneous include call to admin_intro.tpl.php in
wpclef.admin.inc (I presume this call and the file itself can be
re-added later if need be).

git-svn-id: https://plugins.svn.wordpress.org/wpclef/trunk@782165 b8457f37-d9ea-0310-8a92-e5e31aec5664